moved functions not dependent on PDC to common and use then common in rpm helper
