+++ moduleframework/common.py
@@ -871,3 +871,45 @@ def get_helpmd_file(dir_name=conf["docker"]["dockerfiledefaultlocation"]):
                     "envvar to set to another): %s" % helpmdfile)
         helpmdfile = None
     return helpmdfile
+
+
+def getBasePackageSet(modulesDict=None, isModule=True, isContainer=False):
+    """
+    Get list of base packages (for bootstrapping of various module types)
+    It is used internally, you should not use it in case you don't know where to use it.
+
+    :param modulesDict: dictionary of dependent modules
+    :param isModule: bool is module
+    :param isContainer: bool is contaner?
+    :return: list of packages to install
+    """
+    # nspawn container need to install also systemd to be able to boot
+    out = []
+
+    if isModule:
+        if isContainer:
+            out = conf["packages"]["container"]
+        else:
+
+            out = conf["packages"]["common"] + conf["packages"]["basic_modular"]
+    else:
+        if isContainer:
+            out = []
+        else:
+            out = conf["packages"]["common"] + conf["packages"]["basic"]
+    core.print_info("Base packages to install:", out)
+    return out
+
+
+def get_base_compose():
+    default_release = "27"
+    release = os.environ.get("MTF_FEDORA_RELEASE") or default_release
+    if release == "master":
+        release = default_release
+    compose_url = os.environ.get("MTF_COMPOSE_BASE")
+    if not compose_url:
+        tmpcompose = conf.get("compose", {}).get("baseurlrepo")
+        if tmpcompose:
+            compose_url = tmpcompose.format(
+                RELEASE=release, ARCH=conf["generic"]["arch"])
+    return compose_url
